Make DNSMasq only reload if static address table is changed.

+++ dnsmasq/external-dhcp.sh
@@ -3,6 +3,7 @@
 DIR="/home/$USER/orchestrator/dnsmasq"
 HOST="localhost"
 PORT="8998"
+CHANGED="0"
 ADDRESS="fd39:9709:2766:6555::1"
 SERVER_ID=$(/usr/bin/curl -s -k "https://$HOST:$PORT/dhcp/$ADDRESS")
 if [[ "$SERVER_ID" == 'false' ]]; then
@@ -20,6 +21,7 @@ if [[ "$1" == "add" ]]; then
         EXISTS=$(grep "$2" "$DIR/ext-staticaddr")
         if [[ "$EXISTS" == '' ]]; then
             echo "id:$2,[$3]" >> "$DIR/ext-staticaddr"
+            CHANGED="1"
         fi
     fi
 fi
@@ -40,9 +42,11 @@ if [[ $1 == "del" ]]; then
         EXISTS=$(grep --quiet "$2" "$DIR/ext-staticaddr")
         if $EXITS; then
             sed -i "/$3/d" "$DIR/int-staticaddr"
+            CHANGED="1"
         fi
     fi
 fi
 
-
+if [[ "$CHANGED" == "1" ]]; then
 /usr/bin/ps aux | awk '/[d]nsmasq-external/{print $2}' | xargs kill -HUP
+fi
